get collections by id

+++ src/main/java/com/example/Project3/CollectionsController.java
@@ -1,6 +1,7 @@
 package com.example.Project3;
 
 import com.example.Project3.entities.Collections;
+import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,9 +23,10 @@ public class CollectionsController {
     return collectionsRepo.findAll();
   }
   @GetMapping("/{id}")
-  ResponseEntity<Collections> getCollection(@PathVariable long id) {
+  ResponseEntity<Collections> getCollection(@PathVariable Integer id) {
     return collectionsRepo.findById(id)
         .map(ResponseEntity::ok)
         .orElse(ResponseEntity.notFound().build());
   }
+
 }
